cli: move the value formatting code to `clisqlclient`
Release note: None
