+++ pkg/cli/BUILD.bazel
@@ -47,7 +47,6 @@ go_library(
         "quit.go",
         "sql.go",
         "sql_format_table.go",
-        "sql_format_value.go",
         "sql_util.go",
         "sqlfmt.go",
         "start.go",
@@ -99,6 +98,7 @@ go_library(
         "//pkg/ccl/sqlproxyccl",
         "//pkg/ccl/sqlproxyccl/tenantdirsvr",
         "//pkg/cli/cliflags",
+        "//pkg/cli/clisqlclient",
         "//pkg/cli/exit",
         "//pkg/cli/syncbench",
         "//pkg/clusterversion",
@@ -147,7 +147,6 @@ go_library(
         "//pkg/sql/row",
         "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondatapb",
         "//pkg/sqlmigrations",
         "//pkg/storage",
         "//pkg/storage/cloud",
@@ -302,7 +301,6 @@ go_test(
         "nodelocal_test.go",
         "quit_test.go",
         "sql_format_table_test.go",
-        "sql_format_value_test.go",
         "sql_test.go",
         "sql_util_test.go",
         "sqlfmt_test.go",
