+++ pkg/cli/sql_util.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/cockroachdb/cockroach-go/crdb"
 	"github.com/cockroachdb/cockroach/pkg/build"
+	"github.com/cockroachdb/cockroach/pkg/cli/clisqlclient"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catconstants"
@@ -425,12 +426,12 @@ func (c *sqlConn) getLastQueryStatistics() (
 			return 0, 0, 0, 0, 0, containsJobLat, err
 		}
 
-		parseLatencyRaw = formatVal(row[0], iter.colTypes[0], false, false)
-		planLatencyRaw = formatVal(row[1], iter.colTypes[1], false, false)
-		execLatencyRaw = formatVal(row[2], iter.colTypes[2], false, false)
-		serviceLatencyRaw = formatVal(row[3], iter.colTypes[3], false, false)
+		parseLatencyRaw = clisqlclient.FormatVal(row[0], iter.colTypes[0], false, false)
+		planLatencyRaw = clisqlclient.FormatVal(row[1], iter.colTypes[1], false, false)
+		execLatencyRaw = clisqlclient.FormatVal(row[2], iter.colTypes[2], false, false)
+		serviceLatencyRaw = clisqlclient.FormatVal(row[3], iter.colTypes[3], false, false)
 		if containsJobLat {
-			jobsLatencyRaw = formatVal(row[4], iter.colTypes[4], false, false)
+			jobsLatencyRaw = clisqlclient.FormatVal(row[4], iter.colTypes[4], false, false)
 		}
 
 		nRows++
@@ -1075,7 +1076,7 @@ func getColumnStrings(rows *sqlRows, showMoreChars bool) []string {
 	srcCols := rows.Columns()
 	cols := make([]string, len(srcCols))
 	for i, c := range srcCols {
-		cols[i] = formatVal(c, "NAME", showMoreChars, showMoreChars)
+		cols[i] = clisqlclient.FormatVal(c, "NAME", showMoreChars, showMoreChars)
 	}
 	return cols
 }
@@ -1114,7 +1115,7 @@ func getNextRowStrings(rows *sqlRows, colTypes []string, showMoreChars bool) ([]
 
 	rowStrings := make([]string, len(cols))
 	for i, v := range vals {
-		rowStrings[i] = formatVal(v, colTypes[i], showMoreChars, showMoreChars)
+		rowStrings[i] = clisqlclient.FormatVal(v, colTypes[i], showMoreChars, showMoreChars)
 	}
 	return rowStrings, nil
 }
